+++ src/map/utils.rs
@@ -1,11 +1,14 @@
 //! This module contains utilities functions.
 use bevy::prelude::*;
 use bevy_ecs_tilemap::prelude::*;
-use tiled::Map;
+use tiled::{ChunkData, LayerTile, LayerTileData, Map, TileLayer};
 
-/// Convert a [tiled::Map]'s [tiled::Orientation] to a [TilemapType]
+use super::TiledMap;
+
+/// Convert a [Map]'s [tiled::Orientation] to a [TilemapType]
 pub fn get_map_type(map: &Map) -> TilemapType {
     match map.orientation {
+        tiled::Orientation::Orthogonal => TilemapType::Square,
         tiled::Orientation::Hexagonal => match map.stagger_axis {
             tiled::StaggerAxis::X if map.stagger_index == tiled::StaggerIndex::Even => {
                 TilemapType::Hexagon(HexCoordSystem::ColumnOdd)
@@ -23,22 +26,12 @@ pub fn get_map_type(map: &Map) -> TilemapType {
         },
         tiled::Orientation::Isometric => TilemapType::Isometric(IsoCoordSystem::Diamond),
         tiled::Orientation::Staggered => {
-            warn!("Isometric (Staggered) map is not supported");
-            TilemapType::Isometric(IsoCoordSystem::Staggered)
+            panic!("Isometric (Staggered) map is not supported");
         }
-        tiled::Orientation::Orthogonal => TilemapType::Square,
     }
 }
 
-/// Convert a [tiled::Map]'s size to a [TilemapSize]
-pub fn get_map_size(map: &Map) -> TilemapSize {
-    TilemapSize {
-        x: map.width,
-        y: map.height,
-    }
-}
-
-/// Convert a [tiled::Map]'s grid size to a [TilemapGridSize]
+/// Convert a [Map]'s grid size to a [TilemapGridSize]
 pub fn get_grid_size(map: &Map) -> TilemapGridSize {
     TilemapGridSize {
         x: map.tile_width as f32,
@@ -46,95 +39,148 @@ pub fn get_grid_size(map: &Map) -> TilemapGridSize {
     }
 }
 
-/// Convert from Tiled coordinates to a Bevy position.
-///
-/// This function will convert provided Tiled raw position to a Bevy position, according to various maps settings.
-///
-/// Example:
-/// ```rust,no_run
-/// use bevy::prelude::*;
-/// use bevy_ecs_tiled::prelude::*;
-/// use bevy_ecs_tilemap::prelude::*;
-///
-/// let tiled_position = Vec2::new(0., 12.);
-/// let bevy_position = from_tiled_coords_to_bevy(
-///     tiled_position,
-///     &TilemapType::Square,
-///     &TilemapSize::new(0, 3),
-///     &TilemapGridSize::new(16., 16.),
-/// );
-/// ```
-pub fn from_tiled_coords_to_bevy(
-    tiled_position: Vec2,
-    map_type: &TilemapType,
-    map_size: &TilemapSize,
-    grid_size: &TilemapGridSize,
-) -> Vec2 {
-    match map_type {
-        TilemapType::Hexagon(HexCoordSystem::ColumnOdd) => Vec2::new(
-            tiled_position.x + grid_size.x / 4.,
-            (map_size.y as f32 + 0.5) * grid_size.y - tiled_position.y,
-        ),
-        TilemapType::Hexagon(HexCoordSystem::ColumnEven) => Vec2::new(
-            tiled_position.x + grid_size.x / 4.,
-            (map_size.y as f32 + 0.) * grid_size.y - tiled_position.y,
-        ),
-        TilemapType::Hexagon(HexCoordSystem::RowOdd) => Vec2::new(
-            tiled_position.x,
-            map_size.y as f32 * grid_size.y * 0.75 + grid_size.y / 4. - tiled_position.y,
-        ),
-        TilemapType::Hexagon(HexCoordSystem::RowEven) => Vec2::new(
-            tiled_position.x - grid_size.x / 2.,
-            map_size.y as f32 * grid_size.y * 0.75 + grid_size.y / 4. - tiled_position.y,
-        ),
-        TilemapType::Isometric(coords_system) => {
-            from_isometric_coords_to_bevy(tiled_position, coords_system, map_size, grid_size)
+/// Convert a position from Tiled space to Bevy space.
+pub fn from_tiled_coords_to_bevy(tiled_map: &TiledMap, tiled_position: Vec2) -> Vec2 {
+    let map_size = tiled_map.tilemap_size;
+    let map_height = tiled_map.bounding_size.y;
+    let grid_size = get_grid_size(&tiled_map.map);
+    match get_map_type(&tiled_map.map) {
+        TilemapType::Square => {
+            tiled_map.origin_offset
+                + Vec2 {
+                    x: tiled_position.x,
+                    y: map_height - tiled_position.y,
+                }
         }
-        _ => Vec2::new(
-            tiled_position.x,
-            map_size.y as f32 * grid_size.y - tiled_position.y,
-        ),
+        TilemapType::Hexagon(HexCoordSystem::ColumnOdd) => {
+            tiled_map.origin_offset
+                + Vec2 {
+                    x: tiled_position.x,
+                    y: map_height + grid_size.y / 2. - tiled_position.y,
+                }
+        }
+        TilemapType::Hexagon(HexCoordSystem::ColumnEven) => {
+            tiled_map.origin_offset
+                + Vec2 {
+                    x: tiled_position.x,
+                    y: map_height - tiled_position.y,
+                }
+        }
+        TilemapType::Hexagon(HexCoordSystem::RowOdd) => {
+            tiled_map.origin_offset
+                + Vec2 {
+                    x: tiled_position.x,
+                    y: map_height + grid_size.y / 4. - tiled_position.y,
+                }
+        }
+        TilemapType::Hexagon(HexCoordSystem::RowEven) => {
+            tiled_map.origin_offset
+                + Vec2 {
+                    x: tiled_position.x - grid_size.x / 2.,
+                    y: map_height + grid_size.y / 4. - tiled_position.y,
+                }
+        }
+        TilemapType::Isometric(IsoCoordSystem::Diamond) => {
+            let position = iso_projection(
+                tiled_position + tiled_map.origin_offset,
+                &map_size,
+                &grid_size,
+            );
+            Vec2 {
+                x: position.x,
+                y: map_height - grid_size.y / 2. - position.y,
+            }
+        }
+        TilemapType::Isometric(IsoCoordSystem::Staggered) => {
+            panic!("Isometric (Staggered) map is not supported");
+        }
+        _ => unreachable!(),
     }
 }
 
-/// Convert from Tiled isometric coordinates to a Bevy position.
-///
-/// This function will convert provided Tiled raw isometric position to a Bevy position, according to various maps settings.
-///
-/// Example:
-/// ```rust,no_run
-/// use bevy::prelude::*;
-/// use bevy_ecs_tiled::prelude::*;
-/// use bevy_ecs_tilemap::prelude::*;
-///
-/// let tiled_position = Vec2::new(0., 12.);
-/// let bevy_position = from_isometric_coords_to_bevy(
-///     tiled_position,
-///     &IsoCoordSystem::Diamond,
-///     &TilemapSize::new(0, 3),
-///     &TilemapGridSize::new(16., 16.),
-/// );
-/// ```
-pub fn from_isometric_coords_to_bevy(
-    tiled_position: Vec2,
-    iso_coords: &IsoCoordSystem,
-    map_size: &TilemapSize,
+/// Iterate over all tiles from the given [TileLayer]
+pub fn for_each_tile<'a, F>(tiled_map: &'a TiledMap, tiles_layer: &TileLayer<'a>, mut f: F)
+where
+    F: FnMut(LayerTile<'a>, &LayerTileData, TilePos, IVec2),
+{
+    let tilemap_size = tiled_map.tilemap_size;
+    match tiles_layer {
+        TileLayer::Finite(layer) => {
+            for x in 0..tilemap_size.x {
+                for y in 0..tilemap_size.y {
+                    // Transform TMX coords into bevy coords.
+                    let mapped_y = tilemap_size.y - 1 - y;
+                    let mapped_x = x as i32;
+                    let mapped_y = mapped_y as i32;
+
+                    let Some(layer_tile) = layer.get_tile(mapped_x, mapped_y) else {
+                        continue;
+                    };
+                    let Some(layer_tile_data) = layer.get_tile_data(mapped_x, mapped_y) else {
+                        continue;
+                    };
+
+                    f(
+                        layer_tile,
+                        layer_tile_data,
+                        TilePos::new(x, y),
+                        IVec2::new(mapped_x, mapped_y),
+                    );
+                }
+            }
+        }
+        TileLayer::Infinite(layer) => {
+            for (chunk_pos, chunk) in layer.chunks() {
+                // bevy_ecs_tilemap doesn't support negative tile coordinates, so shift all chunks
+                // such that the top-left chunk is at (0, 0).
+                let chunk_pos_mapped = (
+                    chunk_pos.0 - tiled_map.topleft_chunk.0,
+                    chunk_pos.1 - tiled_map.topleft_chunk.1,
+                );
+
+                for x in 0..ChunkData::WIDTH {
+                    for y in 0..ChunkData::HEIGHT {
+                        // Invert y to match bevy coordinates.
+                        let Some(layer_tile) = chunk.get_tile(x as i32, y as i32) else {
+                            continue;
+                        };
+                        let Some(layer_tile_data) = chunk.get_tile_data(x as i32, y as i32) else {
+                            continue;
+                        };
+
+                        let index = IVec2 {
+                            x: chunk_pos_mapped.0 * ChunkData::WIDTH as i32 + x as i32,
+                            y: chunk_pos_mapped.1 * ChunkData::HEIGHT as i32 + y as i32,
+                        };
+
+                        f(
+                            layer_tile,
+                            layer_tile_data,
+                            TilePos {
+                                x: index.x as u32,
+                                y: tilemap_size.y - 1 - index.y as u32,
+                            },
+                            index,
+                        );
+                    }
+                }
+            }
+        }
+    }
+}
+
+/// Convert Tiled isometric coordinates into scalar coordinates
+pub(crate) fn iso_projection(
+    coords: Vec2,
+    tilemap_size: &TilemapSize,
     grid_size: &TilemapGridSize,
 ) -> Vec2 {
-    match iso_coords {
-        IsoCoordSystem::Diamond => Vec2::new(
-            ((tiled_position.x - tiled_position.y) / grid_size.y + map_size.y as f32) * grid_size.x
-                / 2.,
-            (map_size.y as f32
-                - tiled_position.x / grid_size.y
-                - tiled_position.y / grid_size.y
-                - 1.)
-                * grid_size.y
-                / 2.,
-        ),
-        IsoCoordSystem::Staggered => {
-            //warn!("Isometric (Staggered) map is not supported");
-            tiled_position
-        }
+    let fract = Vec2 {
+        x: coords.x / grid_size.y,
+        y: coords.y / grid_size.y,
+    };
+    Vec2 {
+        x: tilemap_size.y as f32 * grid_size.x / 2. + (fract.x - fract.y) * grid_size.x / 2.,
+        y: (fract.x + fract.y) * grid_size.y / 2.,
     }
 }